refactor: merge follows, likes, agent_pins, and ping zomes into mews
Consolidates 5 integrity zomes and 6 coordinator zomes down to
2 integrity (profiles_integrity, mews_integrity) and 2 coordinator
(profiles, mews) zomes. This eliminates cross-zome serialization
overhead and reduces WASM module count from 11 to 4.

- Merge follows/likes/agent_pins link types into mews_integrity LinkTypes enum
- Move coordinator logic into mews, replacing call_local_zome() with direct calls
- Merge ping health check into mews coordinator
- Update hardcoded zome indices in agent_to_notifications.rs
- Update all UI and test zome_name references
- Remove follows_types and hc_call_utils crates

DNA-breaking change: incompatible with networks running the old DNA.
