+++ Cargo.toml
@@ -10,26 +10,6 @@ paste = "1.0"
 hc_prefix_index = "0.14.0"
 holochain_serialized_bytes = "0.0.56"
 
-[workspace.dependencies.agent_pins]
-path = "dnas/mewsfeed/zomes/coordinator/agent_pins"
-
-[workspace.dependencies.agent_pins_integrity]
-path = "dnas/mewsfeed/zomes/integrity/agent_pins"
-
-[workspace.dependencies.follows]
-path = "dnas/mewsfeed/zomes/coordinator/follows"
-
-[workspace.dependencies.follows_integrity]
-path = "dnas/mewsfeed/zomes/integrity/follows"
-
-[workspace.dependencies.follows_types]
-path = "crates/follows_types"
-
-[workspace.dependencies.likes]
-path = "dnas/mewsfeed/zomes/coordinator/likes"
-
-[workspace.dependencies.likes_integrity]
-path = "dnas/mewsfeed/zomes/integrity/likes"
 
 [workspace.dependencies.mews]
 path = "dnas/mewsfeed/zomes/coordinator/mews"
@@ -40,17 +20,12 @@ path = "dnas/mewsfeed/zomes/integrity/mews"
 [workspace.dependencies.mews_types]
 path = "crates/mews_types"
 
-[workspace.dependencies.ping]
-path = "dnas/mewsfeed/zomes/coordinator/ping"
 
 [workspace.dependencies.profiles]
 path = "dnas/mewsfeed/zomes/coordinator/profiles"
 
 [workspace.dependencies.profiles_integrity]
 path = "dnas/mewsfeed/zomes/integrity/profiles"
-
-[workspace.dependencies.hc_call_utils]
-path = "crates/hc_call_utils"
 
 [workspace.dependencies.hc_link_pagination]
 path = "crates/hc_link_pagination"
