--- crates/follows_types/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "follows_types"
-version = "0.0.1"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-name = "follows_types"
-
-[dependencies]
-hdk = { workspace = true }
-serde = { workspace = true }
-hc_link_pagination = { workspace = true }
-holochain_serialized_bytes = { workspace = true }
