--- crates/hc_call_utils/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "hc_call_utils"
-version = "0.0.1"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-name = "hc_call_utils"
-
-[dependencies]
-hdk = { workspace = true }
-serde = { workspace = true }
