+++ dnas/mewsfeed/workdir/dna.yaml
@@ -15,18 +15,6 @@ integrity:
       hash: ~
       path: "../../../target/wasm32-unknown-unknown/release/mews_integrity.wasm"
       dependencies: ~
-    - name: follows_integrity
-      hash: ~
-      path: "../../../target/wasm32-unknown-unknown/release/follows_integrity.wasm"
-      dependencies: ~
-    - name: likes_integrity
-      hash: ~
-      path: "../../../target/wasm32-unknown-unknown/release/likes_integrity.wasm"
-      dependencies: ~
-    - name: agent_pins_integrity
-      hash: ~
-      path: "../../../target/wasm32-unknown-unknown/release/agent_pins_integrity.wasm"
-      dependencies: ~
 coordinator:
   zomes:
     - name: profiles
@@ -40,22 +28,3 @@ coordinator:
       dependencies:
         - name: mews_integrity
         - name: profiles_integrity
-    - name: follows
-      hash: ~
-      path: "../../../target/wasm32-unknown-unknown/release/follows.wasm"
-      dependencies:
-        - name: follows_integrity
-    - name: likes
-      hash: ~
-      path: "../../../target/wasm32-unknown-unknown/release/likes.wasm"
-      dependencies:
-        - name: likes_integrity
-    - name: agent_pins
-      hash: ~
-      path: "../../../target/wasm32-unknown-unknown/release/agent_pins.wasm"
-      dependencies:
-        - name: agent_pins_integrity
-    - name: ping
-      hash: ~
-      path: "../../../target/wasm32-unknown-unknown/release/ping.wasm"
-      dependencies: []
