--- dnas/mewsfeed/zomes/coordinator/agent_pins/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "agent_pins"
-version = "0.0.1"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-name = "agent_pins"
-
-[dependencies]
-hdk = { workspace = true }
-
-serde = { workspace = true }
-
-agent_pins_integrity = { workspace = true }
-hc_zome_input = { workspace = true }
