--- dnas/mewsfeed/zomes/coordinator/agent_pins/src/lib.rs
@@ -1,6 +1,0 @@
-pub mod pinner_to_hashes;
-use hdk::prelude::*;
-#[hdk_extern]
-pub fn init(_: ()) -> ExternResult<InitCallbackResult> {
-    Ok(InitCallbackResult::Pass)
-}
