--- dnas/mewsfeed/zomes/coordinator/follows/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "follows"
-version = "0.0.1"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-name = "follows"
-
-[dependencies]
-hdk = { workspace = true }
-serde = { workspace = true }
-follows_integrity = { workspace = true } 
-hc_link_pagination = { workspace = true }
-hc_zome_input = { workspace = true }
-follows_types = { workspace = true }
