--- dnas/mewsfeed/zomes/coordinator/follows/src/lib.rs
@@ -1,7 +1,0 @@
-use hdk::prelude::*;
-pub mod follower_to_creators;
-
-#[hdk_extern]
-pub fn init(_: ()) -> ExternResult<InitCallbackResult> {
-    Ok(InitCallbackResult::Pass)
-}
