--- dnas/mewsfeed/zomes/coordinator/likes/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "likes"
-version = "0.0.1"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-name = "likes"
-
-[dependencies]
-hdk = { workspace = true }
-
-serde = { workspace = true }
-
-likes_integrity = { workspace = true }
-hc_zome_input = { workspace = true }
