--- dnas/mewsfeed/zomes/coordinator/likes/src/lib.rs
@@ -1,7 +1,0 @@
-pub mod liker_to_hashes;
-use hdk::prelude::*;
-
-#[hdk_extern]
-pub fn init(_: ()) -> ExternResult<InitCallbackResult> {
-    Ok(InitCallbackResult::Pass)
-}
