+++ dnas/mewsfeed/zomes/coordinator/mews/src/licker_to_mews.rs
@@ -1,22 +1,21 @@
-use hc_call_utils::call_local_zome;
+use crate::liker_to_hashes::{count_likers_for_hash, is_liker_for_hash, IsLikerForHashInput};
 use hdk::prelude::*;
+use hc_zome_input::ZomeFnInput;
 
 #[hdk_extern]
 pub fn get_lickers_for_mew(mew_hash: ActionHash) -> ExternResult<Vec<AgentPubKey>> {
-    call_local_zome::<Vec<AgentPubKey>, ActionHash>("likes", "get_likers_for_hash", mew_hash)
+    crate::liker_to_hashes::get_likers_for_hash(ZomeFnInput::new(
+        AnyLinkableHash::from(mew_hash),
+        Some(true),
+    ))
 }
 
 #[hdk_extern]
 pub fn count_lickers_for_mew(mew_hash: ActionHash) -> ExternResult<usize> {
-    call_local_zome::<usize, ActionHash>("likes", "count_likers_for_hash", mew_hash)
+    count_likers_for_hash(AnyLinkableHash::from(mew_hash))
 }
 
-#[derive(Serialize, Deserialize, Debug)]
-pub struct IsLikerForHashInput {
-    pub liker: AgentPubKey,
-    pub hash: AnyLinkableHash,
-}
 #[hdk_extern]
 pub fn is_licker_for_mew(input: IsLikerForHashInput) -> ExternResult<bool> {
-    call_local_zome::<bool, IsLikerForHashInput>("likes", "is_liker_for_hash", input)
+    is_liker_for_hash(input)
 }
