--- dnas/mewsfeed/zomes/coordinator/ping/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "ping"
-version = "0.0.1"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-name = "ping"
-
-[dependencies]
-hdk = { workspace = true }
-serde = { workspace = true }
