--- dnas/mewsfeed/zomes/coordinator/ping/src/lib.rs
@@ -1,6 +1,0 @@
-use hdk::prelude::*;
-
-#[hdk_extern]
-pub fn ping(_: ()) -> ExternResult<()> {
-    Ok(())
-}
