--- dnas/mewsfeed/zomes/integrity/agent_pins/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "agent_pins_integrity"
-version = "0.0.1"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-name = "agent_pins_integrity"
-
-[dependencies]
-hdi = { workspace = true }
-
-serde = { workspace = true }
