--- dnas/mewsfeed/zomes/integrity/follows/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "follows_integrity"
-version = "0.0.1"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-name = "follows_integrity"
-
-[dependencies]
-hdi = { workspace = true }
-
-serde = { workspace = true }
