--- dnas/mewsfeed/zomes/integrity/likes/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "likes_integrity"
-version = "0.0.1"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-name = "likes_integrity"
-
-[dependencies]
-hdi = { workspace = true }
-
-serde = { workspace = true }
