+++ ui/src/components/FollowersListDialog.vue
@@ -73,7 +73,7 @@ const pageLimit = 10;
 const fetchFollowers = async (params: any) => {
   const agents: AgentPubKey[] = await client.callZome({
     role_name: "mewsfeed",
-    zome_name: "follows",
+    zome_name: "mews",
     fn_name: "get_followers_for_creator",
     payload: wrapInput({
       creator: props.agentPubKey,
